Update game protocol to 13.16

+++ src/player.h
@@ -1003,7 +1003,7 @@ class Player final : public Creature, public Cylinder
 				client->sendCreatureShield(creature);
 			}
 		}
-		void sendSpellCooldown(uint8_t spellId, uint32_t time) {
+		void sendSpellCooldown(uint16_t spellId, uint32_t time) {
 			if (client) {
 				client->sendSpellCooldown(spellId, time);
 			}
